requirements.txt
# EvoVLA Dependencies

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0

# OpenVLA related
openvla>=0.1.0
peft>=0.7.0

# Vision
Pillow>=9.0.0
opencv-python>=4.8.0
clip @ git+https://github.com/openai/CLIP.git

# Configuration
omegaconf>=2.3.0
pyyaml>=6.0

# Data processing
numpy>=1.24.0
scipy>=1.10.0
av>=10.0.0  # Video reading

# Visualization and logging
tensorboard>=2.14.0
wandb>=0.16.0
tqdm>=4.65.0

# Optional: DeepSeek API for triplet generation
openai>=1.0.0

# Optional: DISCOVERSE simulator dependencies
# mujoco>=3.0.0
# dm_control>=1.0.0

